+++ spark-on-angel/mllib/src/main/scala/com/tencent/angel/spark/ml/util/package.scala
@@ -108,7 +108,7 @@ package object util {
 
   val ONE_HOT_INSTANCE_ST = StructType(
     StructField(DFStruct.LABEL, StringType, false) ::
-      StructField(DFStruct.FEATURE, ArrayType(IntegerType, containsNull = false), false) ::
+      StructField(DFStruct.FEATURE, ArrayType(LongType, containsNull = false), false) ::
       Nil)
 
 }
